fix(swagger-codegen): reexport utils
