+++ packages/swagger-codegen/src/generate.ts
@@ -1,11 +1,10 @@
 import { generateApi, ParsedRoute, type GenerateApiOutput, type GenerateApiParams } from "swagger-typescript-api";
 import path from "path";
 import { fileURLToPath } from "url";
-import fs from "fs";
+import { createOnCreateRouteMethod, deleteHttpClient, generateIndex, generateModuleForMultiple, generateModuleIfNotExists, generateMultipleIndex, generateUtilsForMultiple } from "./utils";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
-const httpClientFileName = "http-client.ts"
 
 export type GenerateApiOptions = GenerateApiParams & {
     deleteHttpClient?: boolean;
@@ -25,69 +24,6 @@ export interface GenerateApiEndpoint {
 export interface GenerateMultipleApiOptions extends Omit<GenerateApiOptions, 'url'> {
     endpoints: GenerateApiEndpoint[]
 }
-
-const generateIndex = (data: GenerateApiOutput, output: string) => {
-    const { files } = data;
-
-    const imports = files.map(({ name }) => `export * from "./${name.replace(".ts", "")}";`).join("\n");
-
-    try {
-        fs.writeFileSync(path.join(output, "index.ts"), imports + "\n");
-        console.log("Codes are generated \n");
-    }
-    catch (e) {
-        console.log(e);
-        process.exit(1);
-    }
-};
-
-const deleteHttpClient = (data: GenerateApiOutput, output: string) => {
-    const httpClientFile = data.files.find(e => e.name === httpClientFileName);
-
-    if (!httpClientFile) return;
-
-    const filePath = path.join(output, httpClientFile.name);
-
-    if (fs.existsSync(filePath)) {
-        fs.rmSync(filePath);
-        const index = data.files.findIndex(e => e === httpClientFile);
-        data.files.splice(index, 1);
-    }
-}
-
-const generateModuleIfNotExists = () => {
-
-    const moduleDirectory = path.resolve(process.cwd(), "./src/module");
-    const filePath = path.join(moduleDirectory, "index.ts");
-
-    if (fs.existsSync(filePath))
-        return;
-
-
-    const moduleContent =
-        `import { CoreModule, SessionStorageCache } from "@pusula/module.core";
-
-const coreModule = new CoreModule({ key: "CoreModule" });
-coreModule.register(SessionStorageCache, { key: "SessionStorageCache" });
-const injectable = coreModule.createInjectable();
-
-export { coreModule, injectable };
-`;
-
-    if (!fs.existsSync(moduleDirectory)) {
-        fs.mkdirSync(moduleDirectory);
-    }
-
-    try {
-        fs.writeFileSync(path.join(moduleDirectory, "index.ts"), moduleContent);
-
-        console.log("module generated \n");
-    } catch (err) {
-        console.log(err);
-        process.exit(1);
-    }
-
-};
 
 export const generate = async (options: GenerateApiOptions) => {
     const defaultOutput = path.resolve(process.cwd(), "./src/__generated__");
@@ -133,39 +69,6 @@ export const generate = async (options: GenerateApiOptions) => {
     }
 };
 
-const generateMultipleIndex = (endPointNames: { name: string }[], output: string) => {
-
-    const imports = endPointNames.map(({ name }) => `export * from "./${name}";`).join("\n");
-
-    try {
-        fs.writeFileSync(path.join(output, "index.ts"), imports + "\n");
-        console.log("Index generated \n");
-    }
-    catch (e) {
-        console.log(e);
-        process.exit(1);
-    }
-}
-
-const generateModuleForMultiple = (output: string) => {
-    const content = 'export * from "../module";\n';
-
-    try {
-        fs.writeFileSync(path.join(output, "module.ts"), content);
-        console.log("module reexport generated \n");
-    }
-    catch (e) {
-        console.log(e);
-        process.exit(1);
-    }
-}
-
-
-const createOnCreateRouteMethod = (suffix: string) => (routeData: ParsedRoute) => {
-    routeData.raw.moduleName += suffix;
-    routeData.namespace += suffix;
-    return routeData;
-}
 
 export const generateMultiple = async (options: GenerateMultipleApiOptions) => {
     const { endpoints, hooks, ...restOptions } = options;
@@ -191,4 +94,5 @@ export const generateMultiple = async (options: GenerateMultipleApiOptions) => {
 
     generateMultipleIndex(endpoints, output);
     generateModuleForMultiple(output);
+    generateUtilsForMultiple(output);
 }
